+++ src/app/page.tsx
@@ -1,16 +1,5 @@
-"use client";
-
 import { FC } from "react";
-import dynamic from "next/dynamic";
-
-// ThemeTestコンポーネントを動的にインポート
-// isDevelopmentフラグでレンダリングが制御される
-const ThemeTest = dynamic(
-  () => import("@/components/ui/ThemeTest").then((mod) => mod.ThemeTest),
-  {
-    ssr: false,
-  },
-);
+import { DevelopmentOnly } from "@/components/DevelopmentOnly";
 
 type CardProps = {
   title: string;
@@ -56,9 +45,6 @@ const cards: CardProps[] = [
   },
 ];
 
-// ビルド時に評価される環境変数
-const isDevelopment = process.env.NODE_ENV === "development";
-
 export default function Home() {
   return (
     <main className="px-4 py-6 max-w-4xl mx-auto">
@@ -68,33 +54,8 @@ export default function Home() {
         ))}
       </div>
 
-      {/* 開発環境でのみThemeTestコンポーネントをレンダリング */}
-      {isDevelopment && (
-        <>
-          <div className="mt-16 mb-16">
-            <ThemeTest />
-          </div>
-          <section
-            className="mt-16 p-6 rounded-xl shadow-md border border-text/10 transition-all duration-300"
-            style={{ backgroundColor: "var(--card-bg)" }}
-          >
-            <h2 className="text-2xl font-bold text-primary mb-4">
-              UnoCSS + テーマ切り替えのテスト
-            </h2>
-            <p className="mb-4">
-              このセクションは UnoCSS と next-themes
-              を使用したテーマ切り替えの動作確認用です。
-              ヘッダーのテーマトグルボタンをクリックして、ライト/ダークモードを切り替えてみてください。
-            </p>
-            <div className="flex flex-wrap gap-4 mt-6">
-              <button className="btn-primary">プライマリボタン</button>
-              <button className="btn bg-background border border-text/20">
-                デフォルトボタン
-              </button>
-            </div>
-          </section>
-        </>
-      )}
+      {/* 開発環境でのみ表示されるコンポーネント */}
+      <DevelopmentOnly />
     </main>
   );
 }
